Added multi-path LocalFileSystem
You can now pass an array of values to the LocalFileSystem initializer; the paths will be searched sequentially starting from index 0 and return the first file that exists.

This is intended to allow liquid to work with sites which implement view extensions; templates can be over-ridden in extensions by simply including the extension view path before the base view path.